requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=2.1.0=pyhd8ed1ab_0
brotli-python=1.1.0=py39hfa9831e_2
bzip2=1.0.8=h99b78c6_7
ca-certificates=2024.8.30=hf0a4a13_0
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py39h7f933ea_0
charset-normalizer=3.4.0=pyhd8ed1ab_0
etils=1.5.1=pyhd8ed1ab_1
filelock=3.16.1=pyhd8ed1ab_0
freetype=2.12.1=hadb7bae_2
fsspec=2024.10.0=pyhff2d567_0
giflib=5.2.2=h93a5062_0
glfw=3.4=h93a5062_0
gmp=6.3.0=h7bae524_2
gmpy2=2.1.5=py39h0bbb021_2
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=75.1=hfee45f7_0
idna=3.10=pyhd8ed1ab_0
importlib_resources=6.4.5=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
kaldi=5.5.1112=cpu_hbd54a75_5
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libabseil=20240722.0=cxx17_hf9b8971_1
libblas=3.9.0=19_osxarm64_openblas
libcblas=3.9.0=19_osxarm64_openblas
libccd-double=2.1=h9a09cb3_2
libcxx=19.1.3=ha82da77_0
libdeflate=1.22=hd74edd7_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=19_osxarm64_openblas
liblapacke=3.9.0=19_osxarm64_openblas
libmujoco=3.2.4=h75b3d27_1
libopenblas=0.3.24=openmp_hd76b1f2_0
libpng=1.6.44=hc14010f_0
libprotobuf=5.28.2=h8f0b736_0
libsqlite=3.47.0=hbaaea75_1
libtiff=4.7.0=hfce79cd_1
libtorch=2.4.1=cpu_generic_h45d0208_3
libuv=1.49.2=h7ab814d_0
libwebp=1.4.0=h54798ee_0
libwebp-base=1.4.0=h93a5062_0
libxcb=1.17.0=hdb1d25a_0
libzlib=1.3.1=h8359307_2
llvm-openmp=19.1.3=hb52a8e5_0
lodepng=20220109=hf86a087_0
markupsafe=3.0.2=py39h66d85bf_0
mpc=1.3.1=h8f1351a_1
mpfr=4.2.1=hb693164_3
mpmath=1.3.0=pyhd8ed1ab_0
mujoco-python=3.2.4=py39h64b7aae_1
ncurses=6.5=h7bae524_1
networkx=3.2.1=pyhd8ed1ab_0
nomkl=1.0=h5ca1d4c_0
numpy=2.0.2=py39hd1e06cf_0
openfst=1.8.3=h950e60b_3
openjpeg=2.5.2=h9f1df11_0
openssl=3.3.2=h8359307_0
pillow=11.0.0=py39h4ac03e3_0
pip=24.3.1=pyh8b19718_0
pthread-stubs=0.4=hd74edd7_1002
pybind11-abi=4=hd8ed1ab_3
pycparser=2.22=pyhd8ed1ab_0
pyglfw=2.7.0=pyha770c72_1
pyopengl=3.1.6=pyhd8ed1ab_1
pysocks=1.7.1=pyha2e5f31_6
python=3.9.20=h9e33284_1_cpython
python_abi=3.9=5_cp39
pytorch=2.4.1=cpu_generic_py39hf03d321_3
pyyaml=6.0.2=py39h06df861_1
qhull=2020.2=h420ef59_5
readline=8.2=h92ec313_1
requests=2.32.3=pyhd8ed1ab_0
scipy=1.13.1=py39h3d5391c_0
setuptools=75.3.0=pyhd8ed1ab_0
sleef=3.7=h7783ee8_0
sympy=1.13.3=pypyh2585a3b_103
tinyxml2=10.0.0=hebf3989_0
tk=8.6.13=h5083fa2_1
torchaudio=2.4.1=cpu_py39h773889f_1
torchvision=0.19.1=cpu_py39h3a3ee27_1
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024b=hc8b5060_0
urllib3=2.2.3=pyhd8ed1ab_0
wheel=0.44.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hd74edd7_1
xorg-libxdmcp=1.1.5=hd74edd7_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zipp=3.20.2=pyhd8ed1ab_0
zstandard=0.23.0=py39hcf1bb16_1
zstd=1.5.6=hb46c0d2_0
